requirements.txt
# Streamlit Frontend
streamlit>=1.32.0

# Google Generative AI (Gemini)
google-generativeai>=0.5.0

# Text-to-Speech
gTTS==2.5.0

# Environment Variables
python-dotenv==1.0.0

# Utilities
requests==2.31.0
Pillow==10.1.0

# Development (optional)
gunicorn==21.2.0
